Some code tidying, documentation, moving stuff around.
