+++ src/main/kotlin/org/liamjd/bascule/BasculeFileHandler.kt
@@ -133,6 +133,7 @@ class BasculeFileHandler : FileHandler {
 		return source.copyTo(destination, overwrite = true)
 	}
 
+	// TODO: move to the interface
 	fun getFile(folder: File, fileName: String) : File {
 		return File(folder,fileName)
 	}
