+++ src/main/kotlin/org/liamjd/bascule/cache/BasculeCacheImpl.kt
@@ -1,13 +1,18 @@
-package org.liamjd.bascule.scanner
+@file:Suppress("UnusedImport")
+
+package org.liamjd.bascule.cache
 
 import kotlinx.serialization.json.Json
 import kotlinx.serialization.json.JsonConfiguration
-import kotlinx.serialization.set
 import org.liamjd.bascule.lib.FileHandler
 import org.liamjd.bascule.lib.model.Project
 import org.liamjd.bascule.slug
 import java.io.FileNotFoundException
 
+/**
+ * Using kotlinx.serialization to serialize the cache items
+ * The plug-in isn't working correctly in IntelliJ - do not "optimise" the imports
+ */
 class BasculeCacheImpl(val project: Project, val fileHandler: FileHandler) : BasculeCache {
 
 	override fun writeCacheFile(mdCacheItems: Set<MDCacheItem>) {
@@ -25,7 +30,6 @@ class BasculeCacheImpl(val project: Project, val fileHandler: FileHandler) : Bas
 		} catch (fnfe: FileNotFoundException) {
 			return emptySet()
 		}
-
 	}
 
 	private fun getCacheFileName(): String {
