+++ src/main/kotlin/org/liamjd/bascule/cache/LocalDateTimeSerializer.kt
@@ -5,12 +5,6 @@ import kotlinx.serialization.internal.StringDescriptor
 import java.time.LocalDateTime
 import java.time.format.DateTimeFormatter
 
-@Serializable
-data class DocCache(val filePath: String, @Serializable(with = LocalDateTimeSerializer::class) val lastModified: LocalDateTime, val fileSize: Long)
-
-@Serializable
-data class CacheMap(val map: Map<String,DocCache>)
-
 @Serializer(forClass = LocalDateTime::class)
 object LocalDateTimeSerializer: KSerializer<LocalDateTime> {
 	override val descriptor: SerialDescriptor =
