+++ src/main/kotlin/org/liamjd/bascule/cache/MDCacheItem.kt
@@ -1,10 +1,17 @@
-package org.liamjd.bascule.scanner
+package org.liamjd.bascule.cache
 
 import kotlinx.serialization.Serializable
-import org.liamjd.bascule.cache.LocalDateTimeSerializer
 import org.liamjd.bascule.lib.model.PostLink
 import java.time.LocalDateTime
 
+/**
+ * This class is a representation of a source markdown file. Each source markdown file in the project will have a corresponding [MDCacheItem], loaded from a cache file if it exists; otherwise generated on the fly.
+ * The cache item requires:
+ * @param sourceFileSize the length of the markdown file, in bytes
+ * @param sourceFilePath the full filename of the markdown file
+ * @param sourceModificationDate the last modification date of the file, from the filesystem
+ * Note that this class has an unusual override for the [MDCacheItem.equals] function; only the sourceFilePath is considered for equality
+ */
 @Serializable
 class MDCacheItem(val sourceFileSize: Long, val sourceFilePath: String, @Serializable(with = LocalDateTimeSerializer::class) val sourceModificationDate: LocalDateTime) {
