--- src/main/kotlin/org/liamjd/bascule/scanner/BasculeCache.kt
@@ -1,11 +1,0 @@
-package org.liamjd.bascule.scanner
-
-// TODO: move to the library
-interface BasculeCache {
-
-	fun writeCacheFile(mdCacheItems: Set<MDCacheItem>)
-
-	fun loadCacheFile(): Set<MDCacheItem>
-
-}
-
