+++ src/test/kotlin/org/liamjd/bascule/scanner/TEST_DATA.kt
@@ -1,5 +1,6 @@
 package org.liamjd.bascule.scanner
 
+import org.liamjd.bascule.cache.MDCacheItem
 import java.io.File
 import java.time.LocalDate
 import java.time.Month
